[Robot] Remove intake1 and intake2
